Started implementing LevelStates
The level states are meant to handle the different States the game can be in, while it's active. However, this requires rethinking of the displayService or the Game Panel.

Either make it so that the displayService uses multiple panels (one of which is the gamePanel) or make it so that the GamePanel uses different Views.

The former makes more sense on first thought, since the Panels can just be added registered to the displayService.